Only attempt to recreate channel when the connection is still good. Move patching to own file
